feat: micro-optimize priority
We send them as fixed-sized 32bit ints on the wire, so we don't need to store them locally as full ints.

This should bring each Entry down from 32 bytes to 24 bytes because the Priority and WantType should now fit into 8 bytes with no padding.

Is this super important? Probably not. But it is showing up in pprof (although most of the space is probably going to the underlying CIDs).