fix: Nil dereference while using SetSendDontHaves
This option is used by the benchmark to simulate the old bitswap
comportement.

This follows the same refactoring idea as done in 47b99b1ce34a.

It was crashing since it was trying to access the `sendDontHaves`
property of `bs.engine` but `bs.engine` is initialized right after
the options are applied, not before.