feat: add basic tracing (#562)
This adds tracing spans to the costly Bitswap entry points. It doesn't
instrument the bitswap internals, which will take some time. In
go-ipfs, this will at least let us know the contribution of Bitswap to
the overall request handling time.

This also plumbs contexts through internally so that they reach the
content routing APIs, so that traces are propagated through and we can
start instrumenting e.g. the DHT.